feat(pi): wire the secret vault into wizard_ask and set_env_values (#923)
